+++ fj-core/src/test/java/test/org/fugerit/java/core/cli/TestArgUtils.java
@@ -20,6 +20,11 @@ public class TestArgUtils {
 			ArgUtils.getArgString( "input" ), "test-input",
 	};
 	
+	private static final String[] ARGS_ALT = {
+			ArgUtils.getArgString( ArgUtils.ARG_PARAM_FILE ), "src/test/resources/core/cli/param1.properties",
+			ArgUtils.getArgString( "force" ),
+	};
+	
 	@Test
 	public void test1() {
 		Properties params = ArgUtils.getArgs( ARGS, true, false );
@@ -44,6 +49,21 @@ public class TestArgUtils {
 	@Test
 	public void test4() {
 		Properties params = ArgUtils.getArgs( ARGS, false, false );
+		log.info( "params : {}", params );
+		Assert.assertNull( params.getProperty( TEST_PARAM_NAME ) );
+	}
+	
+	@Test
+	public void testAlt() {
+		Properties params = ArgUtils.getArgs( ARGS_ALT, false, false );
+		log.info( "params : {}", params );
+		Assert.assertNull( params.getProperty( TEST_PARAM_NAME ) );
+	}
+
+	@Test
+	public void testNoParam() {
+		String[] args = { "test" };
+		Properties params = ArgUtils.getArgs( args, false, false );
 		log.info( "params : {}", params );
 		Assert.assertNull( params.getProperty( TEST_PARAM_NAME ) );
 	}