+++ fj-core/src/test/java/test/org/fugerit/java/core/lang/annotate/TestImplFinder.java
@@ -1,0 +1,5 @@
+package test.org.fugerit.java.core.lang.annotate;
+
+public class TestImplFinder {
+
+}