+++ fj-core/src/test/java/test/org/fugerit/java/core/lang/binding/ModelTestBindingTwo.java
@@ -1,0 +1,12 @@
+package test.org.fugerit.java.core.lang.binding;
+
+import lombok.Data;
+
+@Data
+public class ModelTestBindingTwo {
+
+	private String fieldTwo;
+	
+	private ModelTestHelper testHelperTwo;
+	
+}