+++ fj-core/src/test/java/test/org/fugerit/java/core/lang/binding/ModelTestHelper.java
@@ -1,0 +1,12 @@
+package test.org.fugerit.java.core.lang.binding;
+
+import org.fugerit.java.core.lang.annotate.DefineImpl;
+
+@DefineImpl( as = ModelTestHelperImpl.class )
+public interface ModelTestHelper {
+
+	String getMyField();
+
+	void setMyField(String myField);
+
+}